requirements.txt
streamlit==1.31.0
fpdf==1.7.2
pillow==10.2.0
